master - update sdk

+++ component/soc/amebalite/fwlib/ram_common/ameba_gdma_memcpy.c
@@ -35,7 +35,7 @@ TIMESENSITIVE_TEXT_SECTION
 void memcpy_gdma_init(void)
 {
 	gdma_memcpy.dma_done = 1;
-	gdma_memcpy.ch_num = GDMA_ChnlAlloc(0, (IRQ_FUN)memcpy_gdma_int, NULL, INT_PRI5);
+	gdma_memcpy.ch_num = GDMA_ChnlAlloc(0, (IRQ_FUN)memcpy_gdma_int, (u32)NULL, INT_PRI5);
 
 	GDMA_StructInit(&(gdma_memcpy.GDMA_InitStruct));
 	gdma_memcpy.GDMA_InitStruct.GDMA_ChNum = gdma_memcpy.ch_num;
